+++ src/discord/logger.py
@@ -86,29 +86,11 @@ class Logger(commands.Cog):
     async def on_member_join(self, member: discord.Member):
         """
         Executes when a member joins. Completes the following actions:
-            * Sends a welcome message to the #welcome channel.
             * Sends a message to #lounge if the number of members ends with two zeros.
 
         Args:
             member (discord.Member): The member who just joined the server.
         """
-        join_channel = discord.utils.get(
-            member.guild.text_channels,
-            name=CHANNEL_WELCOME,
-        )
-        assert isinstance(join_channel, discord.TextChannel)
-        await join_channel.send(
-            f"{member.mention}, welcome to the Scioly.org Discord Server! "
-            "You can add roles here, using the commands shown at the top of "
-            "this channel. If you have any questions, please just ask here, "
-            "and a helper or moderator will answer you ASAP."
-            "\n\n"
-            "**Please add roles by typing the commands above into the text box,"
-            " and if you have a question, please type it here. After adding "
-            "roles, a moderator will give you access to the rest of the server to "
-            "chat with other members!**",
-        )
-
         # Send fun alert message on every 100 members who join
         member_count = len(member.guild.members)
         lounge_channel = discord.utils.get(
@@ -353,18 +335,22 @@ class Logger(commands.Cog):
         """
         # Get the required resources for logging
         channel = self.bot.get_channel(payload.channel_id)
+        if not channel:
+            channel = await self.bot.fetch_channel(payload.channel_id)
         guild = (
             self.bot.get_guild(SERVER_ID)
-            if channel.type == discord.ChannelType.private
+            if isinstance(channel, discord.abc.PrivateChannel)
             else channel.guild
         )
-        edited_channel: discord.TextChannel = discord.utils.get(
+        if not guild:
+            guild = await self.bot.fetch_guild(SERVER_ID)
+        edited_channel = discord.utils.get(
             guild.text_channels,
             name=CHANNEL_EDITEDM,
         )
 
         # Ignore payloads for events in logging channels (which would cause recursion)
-        if channel.type != discord.ChannelType.private and channel.name in [
+        if not isinstance(channel, discord.abc.PrivateChannel) and channel.name in [
             CHANNEL_EDITEDM,
             CHANNEL_DELETEDM,
             CHANNEL_DMLOG,
@@ -451,7 +437,12 @@ class Logger(commands.Cog):
             await edited_channel.send(embed=embed)
 
         except Exception:  # No cached message is available
-            message_now = await channel.fetch_message(payload.message_id)
+            try:
+                message_now = await channel.fetch_message(payload.message_id)
+            except Exception:
+                # Chances are the message was deleted before we could fetch it
+                # or it comes from an ephemeral response
+                return
             embed = discord.Embed(
                 title=":pencil: Edited Message",
                 color=discord.Color.orange(),